feat(options): add SyncLogger
