+++ package.json
@@ -47,5 +47,7 @@
   "peerDependencies": {
     "almin": ">= 0.3.0"
   },
-  "dependencies": {}
+  "dependencies": {
+    "@azu/format-text": "^1.0.1"
+  }
 }
