+++ src/AlminLogger.js
@@ -1,10 +1,19 @@
 // LICENSE : MIT
 "use strict";
 const EventEmitter = require("events");
-// performance.now polyfill
-import now from "./performance-now";
+const format = require('@azu/format-text');
+import SyncLogger from "./SyncLogger";
+import AsyncLogger from "./AsyncLogger";
+const DefaultTemplates = {
+    WaringMixedUseCaseExecution: "Warning: Executing multiple UseCase at once {useCases}"
+};
 const DefaultOptions = {
-    console: console
+    // output log asynchronously
+    async: true,
+    // use `console` object for logging
+    console: console,
+    // message template object
+    templates: DefaultTemplates
 };
 export default class AlminLogger extends EventEmitter {
     /**
@@ -13,64 +22,33 @@ export default class AlminLogger extends EventEmitter {
      */
     static get Events() {
         return {
-            start: "start",
-            output: "output",
-            flush: "flush",
-            release: "release"
+            output: "output"
         };
     }
 
-    constructor({console} = {}) {
+    constructor(options = {}) {
         super();
-        this._logBuffer = [];
-        this._logMap = {};
-        this._releaseHandlers = [];
         // default logger is `console`
-        this.logger = console || DefaultOptions.console;
-    }
-
-    /**
-     * show current buffer using logger.
-     * @param {string} logTitle
-     */
-    outputBuffer(logTitle) {
-        const output = (log) => {
-            if (log instanceof Error) {
-                this.logger.error(error);
-            } else {
-                this.logger.info(log);
-            }
+        const console = options.console || DefaultOptions.console;
+        const templates = options.templates || DefaultOptions.templates;
+        const isAsyncMode = options.async !== undefined ? options.async : DefaultOptions.async;
+        const loggerOptions = {
+            console,
+            templates
         };
-        this.logger.groupCollapsed(logTitle);
-        // if executing multiple UseCase at once, show warning
-        const currentExecuteUseCases = this._logBuffer.filter(logBuffer=> {
-            return logBuffer.indexOf("will execute") !== -1;
+        /**
+         * @type {boolean} if current is async mode, return true
+         */
+        this.isAsyncMode = isAsyncMode;
+        // default: Async logger
+        this.logger = isAsyncMode ? new AsyncLogger(loggerOptions) : new SyncLogger(loggerOptions);
+        this.logger.on(AlminLogger.Events.output, () => {
+            this.emit(AlminLogger.Events.output);
         });
-        if (currentExecuteUseCases.length > 1) {
-            this.logger.warn("Warning: Executing multiple UseCase at once", currentExecuteUseCases.map(name => {
-                return name.replace(" will execute", "");
-            }))
-        }
-        this._logBuffer.forEach(logBuffer => {
-            if (Array.isArray(logBuffer)) {
-                const title = logBuffer.shift();
-                this.logger.groupCollapsed(title);
-                logBuffer.forEach(output);
-                this.logger.groupEnd();
-            } else {
-                output(logBuffer);
-            }
-        });
-        this.logger.groupEnd();
-        this.emit(AlminLogger.Events.output);
     }
 
-    /**
-     * flush current log buffer
-     */
     flushBuffer() {
-        this._logBuffer.length = 0;
-        this.emit(AlminLogger.Events.flush);
+        this.logger.flushBuffer();
     }
 
     /**
@@ -78,84 +56,13 @@ export default class AlminLogger extends EventEmitter {
      * @param {Context} context
      */
     startLogging(context) {
-        this._logMap = {};
-        this._logBuffer = [];
-        this._releaseHandlers = [];
-        /**
-         * @param {UseCase} useCase
-         */
-        const onWillExecuteEachUseCase = useCase => {
-            this._logMap[useCase.name] = now();
-            this._logBuffer.push(`${useCase.name} will execute`)
-        };
-        const onDispatch = payload => {
-            this._logDispatch(payload);
-        };
-        const onChange = (stores) => {
-            this._logOnChange(stores);
-        };
-        const onErrorHandler = (error) => {
-            this._logError(error);
-        };
-        const onDidExecuteEachUseCase = useCase => {
-            const timeStamp = this._logMap[useCase.name];
-            const takenTime = now() - timeStamp;
-            this._logBuffer.push(`${useCase.name} did executed`);
-            this._logBuffer.push("Taken time(ms): " + takenTime);
-            this.outputBuffer(`\u{1F516} ${useCase.name}`);
-            this.flushBuffer();
-        };
-        // release handler
-        this._releaseHandlers = [
-            context.onChange(onChange),
-            context.onDispatch(onDispatch),
-            context.onWillExecuteEachUseCase(onWillExecuteEachUseCase),
-            context.onDidExecuteEachUseCase(onDidExecuteEachUseCase),
-            context.onErrorDispatch(onErrorHandler)
-        ];
-        this.emit(AlminLogger.Events.start);
-
-    }
-
-    _logError(payload) {
-        // if has useCase and group by useCase
-        if (payload.useCase) {
-            this._logBuffer.push([
-                payload.useCase.name,
-                payload.error
-            ]);
-        } else {
-            this._logBuffer.push(payload.error);
-        }
-    }
-
-    _logDispatch(payload) {
-        // http://emojipedia.org/fire/
-        this._logBuffer.push([
-            `\u{1F525} Dispatch:${payload.type}`,
-            payload
-        ]);
-    }
-
-    /**
-     * @param {Store[]} stores
-     */
-    _logOnChange(stores) {
-        stores.forEach(store => {
-            // http://emojipedia.org/floppy-disk/
-            this._logBuffer.push([
-                `\u{1F4BE} Store:${store.name} is Changed`,
-                store.getState()
-            ]);
-        })
+        this.logger.startLogging(context);
     }
 
     /**
      * release event handlers
      */
     release() {
-        this._releaseHandlers.forEach(releaseHandler => releaseHandler());
-        this._releaseHandlers.length = 0;
-        this.emit(AlminLogger.Events.release);
+        this.logger.release();
     }
 }
