+++ test/AlminLogger-test.js
@@ -1,20 +1,26 @@
 // LICENSE : MIT
 "use strict";
 const assert = require("assert");
-import {Context, Store, UseCase, Dispatcher} from "almin";
+import {Context, Store, Dispatcher} from "almin";
 import AlminLogger from "../src/AlminLogger";
+import AsyncLogger from "../src/AsyncLogger";
+import SyncLogger from "../src/SyncLogger";
 import ConsoleMock from "./helper/ConsoleMock";
-class ExampleUseCase extends UseCase {
-    execute() {
-
-    }
-}
-class DispatchUseCase extends UseCase {
-    execute(payload) {
-        this.dispatch(payload);
-    }
-}
+import ExampleUseCase from "./usecase/ExampleUseCase"
 describe("AlminLogger-test", function () {
+    context("async options", function () {
+        it("use async logger by default", function () {
+            const logger = new AlminLogger();
+            assert(logger.logger instanceof AsyncLogger);
+        });
+        it("when set async:false, use SyncLogger", function () {
+            const logger = new AlminLogger({
+                async: false
+            });
+            assert(logger.logger instanceof SyncLogger);
+        });
+    });
+
     it("should log useCase", function (done) {
         const consoleMock = ConsoleMock.create();
         const logger = new AlminLogger({
@@ -41,62 +47,6 @@ describe("AlminLogger-test", function () {
         dispatcher.dispatch({
             type: "Test"
         });
-        context.useCase(useCase).execute();
-    });
-    it("should log dispatch event", function (done) {
-        const consoleMock = ConsoleMock.create();
-        const logger = new AlminLogger({
-            console: consoleMock
-        });
-        const dispatcher = new Dispatcher();
-        const store = new Store();
-        const context = new Context({
-            store,
-            dispatcher
-        });
-        const useCase = new DispatchUseCase();
-        logger.startLogging(context);
-        // yet not called
-        assert(!consoleMock.groupCollapsed.called);
-        assert(!consoleMock.log.called);
-        // Then
-        logger.on(AlminLogger.Events.output, function () {
-            assert(consoleMock.groupCollapsed.called);
-            const expectOutput = `Dispatch:example`;
-            const isContain = consoleMock.log.calls.some(call => {
-                return call.arg.indexOf(expectOutput) !== -1;
-            });
-            assert(isContain, `${expectOutput} is not found.`);
-            done();
-        });
-        // When
-        context.useCase(useCase).execute({
-            type: "example"
-        });
-    });
-    it("should output as async", function (done) {
-        const consoleMock = ConsoleMock.create();
-        const logger = new AlminLogger({
-            console: consoleMock
-        });
-        const dispatcher = new Dispatcher();
-        const store = new Store();
-        const useCase = new ExampleUseCase();
-        const context = new Context({
-            store,
-            dispatcher
-        });
-        logger.startLogging(context);
-        // yet not called
-        assert(!consoleMock.groupCollapsed.called);
-        assert(!consoleMock.log.called);
-        // Then
-        logger.on(AlminLogger.Events.output, function () {
-            assert(consoleMock.groupCollapsed.called);
-            assert(consoleMock.log.called);
-            done();
-        });
-        // When
         context.useCase(useCase).execute();
     });
 });
