+++ test/usecase/DispatchUseCase.js
@@ -1,0 +1,8 @@
+// LICENSE : MIT
+"use strict";
+import {UseCase} from "almin";
+export default class DispatchUseCase extends UseCase {
+    execute(payload) {
+        this.dispatch(payload);
+    }
+}
