+++ test/usecase/ExampleUseCase.js
@@ -1,0 +1,8 @@
+// LICENSE : MIT
+"use strict";
+import {UseCase} from "almin";
+export default class ExampleUseCase extends UseCase {
+    execute() {
+
+    }
+}
